db:migrate

+++ db/migrate/20190607014221_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[5.1]
   def change
     create_table :products do |t|
       t.string :name
-      t.double :price
+      t.decimal :price
       t.integer :quantity
       t.text :description
       t.references :admin, foreign_key: true
